pyproject.toml
[project]
name = "talk"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.101.0",
    "openai-agents>=0.2.9",
    "pydantic-ai[logfire]>=0.7.4",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.12.10",
]
